requirements.txt
openai
anthropic
google-generativeai
transformers
torch
